Fixed issue with incorrect extension method being called
Umbraco/NPoco provides several "OrderBy" extension methods for "Sql<T>", so tghe compiler would pick an incorrect overload, causing errors when sorting in ascending order. To be more specific about the method that we wish to use, this commit changes the code to call the via "NPocoSqlExtensions.OrderBy" like a static method instead.